vista-iot-gateway-backend/requirements.txt
fastapi>=0.68.0
uvicorn>=0.15.0
pydantic>=1.8.2
pyyaml>=6.0
pymodbus>=3.0.0
pyserial>=3.5
psutil>=5.9.0
aiohttp>=3.8.1
python-multipart>=0.0.5
asyncio>=3.4.3
